Change edge weight type to double

+++ DPSolution.h
@@ -11,7 +11,7 @@ class DPSolution : public ISolution {
     public:
         DPSolution(const MatrixGraph&);
 
-        virtual std::pair<unsigned, std::vector<unsigned> > run() override;
+        virtual std::pair<double, std::vector<unsigned> > run() override;
 };
 
 #endif
