+++ NNSolution.h
@@ -11,7 +11,7 @@ class NNSolution : public ISolution {
     public:
         NNSolution(const MatrixGraph&);
 
-        virtual std::pair<unsigned, std::vector<unsigned> > run() override;
+        virtual std::pair<double, std::vector<unsigned> > run() override;
 };
 
 #endif
